refactor :: Change class to object
