+++ src/main/kotlin/hs/kr/equus/feed/infrastructure/kafka/consumer/reply/CreateReplyConsumer.kt
@@ -1,7 +1,7 @@
 package hs.kr.equus.feed.infrastructure.kafka.consumer.reply
 
 import hs.kr.equus.feed.domain.question.service.MarkAsRepliedService
-import hs.kr.equus.user.infrastructure.kafka.configuration.KafkaTopics
+import hs.kr.equus.feed.infrastructure.kafka.configuration.KafkaTopics
 import hs.kr.equus.user.infrastructure.kafka.dto.CreateReplyEventRequest
 import org.springframework.kafka.annotation.KafkaListener
 import org.springframework.stereotype.Service
