+++ src/main/kotlin/hs/kr/equus/feed/infrastructure/kafka/producer/reply/CreateReplyProducerImpl.kt
@@ -1,6 +1,6 @@
 package hs.kr.equus.feed.infrastructure.kafka.producer.reply
 
-import hs.kr.equus.user.infrastructure.kafka.configuration.KafkaTopics
+import hs.kr.equus.feed.infrastructure.kafka.configuration.KafkaTopics
 import hs.kr.equus.user.infrastructure.kafka.dto.CreateReplyEventRequest
 import org.springframework.kafka.core.KafkaTemplate
 import org.springframework.stereotype.Service
